fixed atoi ret value and ok_input

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: amema <amema@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 18:11:14 by amema             #+#    #+#             */
-/*   Updated: 2023/02/01 18:12:04 by amema            ###   ########.fr       */
+/*   Updated: 2023/10/05 14:39:00 by amema            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int	i;
 	int	sign;
-	int	result;
+	long	result;
 
 	i = 0;
 	while ((str[i] >= 9 && str[i] <= 13) || (str[i] == 32))
